pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wada-compliance-agent"
version = "1.0.0"
description = "Multi-agent WADA compliance checker using Google ADK + Gemini"
requires-python = ">=3.11"
dependencies = [
    # Google AI
    "google-adk>=0.4.0",
    "google-generativeai>=0.8.0",
    "google-cloud-aiplatform>=1.60.0",

    # API framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",

    # HTTP + scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",

    # Vector search
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.3.0",        # retry logic
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",          # for TestClient
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["agents", "api"]
